Fix some nits from previous commit
- move process out of main
- correct some grammar in log msg
